Setup ZFS and tmpfs root
This is by no means a complete filesystem configuration. Persistence
hasn't even been setup yet. This is just enough to boot and define the
sources and mountpoints for the minecraft server state and persistent
datasets. Root is currently mounted on a tmpfs since it's the simplest
implementation of an impermanent root, though I may later setup an
impermanent zfs dataset to allow backups via snapshots.